+++ frontend/lib/event_list_component_po.dart
@@ -2,6 +2,7 @@ import 'dart:async';
 import 'package:angular2/router.dart';
 import 'package:mockito/mockito_no_mirrors.dart';
 import 'package:pageloader/objects.dart';
+import 'package:frontend/store/regatta_state.dart';
 import 'package:frontend/store/regatta_store.dart';
 import 'package:greencat/greencat.dart';
